requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.15.2
psycopg==3.2.3
psycopg-pool==3.2.4
redis==5.2.1
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.8.2
